+++ Hspi/OpenZWaveDB/ZWaveEndPoints.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+#nullable enable
+
+namespace Hspi.OpenZWaveDB
+{
+    internal record ZWaveEndPoints
+    {
+        public IReadOnlyList<ZWaveCommandClass>? CommandClass { get; init; }
+    }
+}
