+++ Hspi/OpenZWaveDB/ZWaveInformation.cs
@@ -7,48 +7,6 @@ using System.Linq;
 
 namespace Hspi.OpenZWaveDB
 {
-    internal record ZWaveEndPoints
-    {
-        public IReadOnlyList<ZWaveCommandClass>? CommandClass { get; init; }
-    }
-
-    internal record ZWaveCommandClass
-    {
-        [JsonProperty("commandclass_name")]
-        public string? Name { get; init; }
-
-        public IReadOnlyList<ZWaveCommandClassChannel>? Channels { get; init; }
-
-        [JsonIgnore]
-        public bool IsSetCommand => Name == "COMMAND_CLASS_CONFIGURATION";
-    }
-
-    internal record ZWaveCommandClassChannel
-    {
-        [JsonProperty("config")]
-        public string? Config { get; init; }
-
-        public string? Label { get; init; }
-        public string? Overview { get; init; }
-
-        [JsonIgnore]
-        public int? ParameterId
-        {
-            get
-            {
-                var list = Config?.Split(new char[] { '=' }, StringSplitOptions.RemoveEmptyEntries);
-                if (list != null && list.Length == 2)
-                {
-                    if (int.TryParse(list[1], out var value))
-                    {
-                        return value;
-                    }
-                }
-                return null;
-            }
-        }
-    }
-
     internal record ZWaveInformation
     {
         [JsonProperty("database_id")]
